Implement keystone bundle integration test and address PR feedback

+++ lib/charms/mysql/v0/mysql.py
@@ -355,7 +355,14 @@ class MySQLBase(ABC):
             raise MySQLCreateApplicationDatabaseAndScopedUserError(e.message)
 
     def delete_users_for_unit(self, unit_name: str) -> None:
-        """Delete users for a unit."""
+        """Delete users for a unit.
+
+        Args:
+            unit_name: The name of the unit for which to delete mysql users for
+
+        Raises:
+            MySQLDeleteUsersForUnitError if there is an error deleting users for the unit
+        """
         get_unit_user_commands = (
             f"SELECT CONCAT(user.user, '@', user.host) FROM mysql.user AS user JOIN information_schema.user_attributes AS attributes ON (user.user = attributes.user AND user.host = attributes.host) WHERE attributes.attribute LIKE '%\"unit_name\": \"{unit_name}\"%'",
         )
@@ -363,6 +370,10 @@ class MySQLBase(ABC):
         try:
             output = self._run_mysqlcli_script("; ".join(get_unit_user_commands))
             users = [line.strip() for line in output.split("\n") if line.strip()][1:]
+
+            if len(users) == 0:
+                logger.debug(f"There are no users to drop for unit {unit_name}")
+                return
 
             drop_users_commands = (f"DROP USER IF EXISTS {', '.join(users)}",)
             self._run_mysqlcli_script("; ".join(drop_users_commands))
